Alex Waugh:
Files need to be specified in o.foo format, otherwise libfile can't find
them. However there is a bug in OS_GBPB.s that meant libfile would never
find any files, even if they existed.

The following patch fixes the bug in OS_GBPB.s, changes all of the SWI
wrappers to return an error pointer, or NULL, gives a more freindly error
message if it can't open the output file, and gives a warning if it can't
find an input file.


git-svn-id: svn://svn.riscos.info/gccsdk/branches/release_2_95@394 5d7976f4-0b0e-0410-a586-d2bac8e3be52