Alex Waugh:
When the x option is specified to ar, it should extract all members
listed on the command line, but if no members are specified, it should
extract all members. The latter wasn't implemented, and the former,
along with the corresponding libfile -e, was broken (the destDir string
was never suitably initialised).

Additionally, the libfile -q option was ignored when -x was specified.


git-svn-id: svn://svn.riscos.info/gccsdk/branches/release_2_95@481 5d7976f4-0b0e-0410-a586-d2bac8e3be52