as fixes from Tony van der Hoff:
Added various comments to the code, to help me understand it.

GET files were not being found via the -I path:
added some debuggging code in as/include.c; found that "." was being
used as directory separator.
- In as/uname.h changed switch from "#ifdef UNIX" to "#if defined
CROSS_COMPILE"

Local labels with default scope were giving warning:
"Assumed 'a' in local label name". This is not a dodgy code warning, but
a warning that the range qualifiers for local labels is not implemented.
- changed lex.c to produce more appropriate warning, and wrapped warning
under "pedantic".

AS barfed on variable substitution operator "$":
- inputUnGet() called from lexGetId() assumes that the input buffer
is pointed to by |input_buff|. It isn't in this case. The check against
the buffer seems superfluous anyway. Fixed by introducing a variant
form of inputUnGet(), imaginitively called inputPutBack().


git-svn-id: svn://svn.riscos.info/gccsdk/trunk@703 5d7976f4-0b0e-0410-a586-d2bac8e3be52