Previous change (checking reserved & unsupported symbol attributes) results in a fatal error when some older AOF files produced by a buggy 'as' version were used.
'as' has been fixed but it might well be that a lot of users
are using this newer drlink version on their older as-produced AOF files.

So therefore we relax the error a bit and downgrade it to a warning and
also report that this might be because of a buggy old 'as' version.


git-svn-id: svn://svn.riscos.info/gccsdk/trunk@1721 5d7976f4-0b0e-0410-a586-d2bac8e3be52