These changes make executables/libraries built in future incompatible with previously built ones. Full rebuild required.
Optimise generated PLTs:

* bfd/elf32-arm.c:
Executable first PLT entry: No change required.

Executable normal PLT entries: Simplify by observing that the public GOT
and the private GOT are one in the same, ie, they have the same address.
It is therefore not necessary to go through the process of accessing
the public GOT and runtime array to find the address of the private GOT.
We can use code similar to ARM Linux, although using r8 instead of r12.

Shared library first PLT entry: Rather than accessing the public GOT and
passing that to the dynamic linker which then has to convert to the private
GOT, take advantage of the fact that the PIC register already has the
private GOT loaded and pass that directly to the dynamic linker instead.

Shared library normal PLT entries: No change required.

* riscos/soloader/dynamic/ld.so-1.9.9/d-link/riscos/resolve.s:
Dynamic linker is passed the client's private GOT pointer directly removing
the need for it to be found using the public version and runtime array.

* riscos/soloader/dynamic/ld.so-1.9.9/d-link/readelflib1.c:
Now that the dynamic linker does not have to access the public GOT to find
the private GOT, it is no longer necessary to write to the public version,
which was notionally a bad thing.

* riscos/soloader/dist/!SharedLibs/docs/Technical: Update docs.
* riscos/dist/!GCC/docs/SharedLibs/!ReadMe: Likewise.
* riscos/dist/!GCC/docs/SharedLibs/Technical: Likewise.


git-svn-id: svn://svn.riscos.info/gccsdk/trunk/gcc4@3506 5d7976f4-0b0e-0410-a586-d2bac8e3be52