* gcc/unwind-dw2.c: * (uw_update_context): Don't use (context->cfa - 4) as the frame pointer, as on rare occasions the offset is -8 instead of -4 for functions that have an extra stack adjustment instruction in their prologue. Instead extract the frame pointer from the context before it's updated.
* (uw_init_context): Don't save the stack limit in the context. We can't
efficiently update it during stack unwinding anyway as it's not stored in
the stack frames. Better to just work it out when we install the context.

* (uw_install_context): It's possible that the stack unwinder, while
stepping through the call frames, may have stepped past those frames
contained within the current stack chunk. When it eventually finds a
context that requires that execution returns to the corresponding frame,
that frame may not be within the current stack chunk leaving the stack
in an undefined state. To bring the stack back to order, we must restore
the stack limit register for the chunk we end up in and free any chunks
that have been left dangling.
Use __ehs_stack_limit(), with the stack pointer extracted from the
current context, to find the stack limit and restore it.
__ehs_trim_stack() is a wrapper function for __trim_stack that preserves
any registers that __trim_stack may use. Instead of jumping straight to
the handler, the handler's address is stored on the stack of the context
that is to be installed and the address of __ehs_trim_stack() is
substituted. When __builtin_eh_return is called, all registers of the
caller are restored and __ehs_trim_stack() is called as if it was the
handler. At this pointer sp points to the real handler's address (as
stored on the stack earlier) which __ehs_trim_stack() can retrieve.
It calls the __trim_stack function to free any unused stack chunks and
then returns to the real exception handler.


git-svn-id: svn://svn.riscos.info/gccsdk/trunk/gcc4@3838 5d7976f4-0b0e-0410-a586-d2bac8e3be52