Fix cause of stack corruption first seen in Vala.
arm_leaf_function_p() caches the result of leaf_function_p() on first
call and returns the cached result on subsequent calls. This works
99.99% of the time, but sometimes first impressions are wrong as in
this case. As GCC processes a function, it's possible for the leaf
status of the function to change making the cached result invalid.
This can lead to a non-leaf function that uses stack space having
no stack checking code in its prologue. If a number of these functions
call each other, as in Vala's case, the stack pointer can pass the
stack limit and overwrite the stack chunk header giving a corrupt
stack.
We could just call leaf_function_p() directly and forget
arm_leaf_function_p() as newer version of GCC appear to do, but tests
show that at the point where the need for stack checking is determined,
the wrong result is returned.
Instead, a more conservative approach is used whereby if GCC has
initially determined that a function is a leaf, it can change it to
non-leaf at a future point, but once it becomes non-leaf, it stays
non-leaf.
This may lead to leaf functions with unnecessary stack checking, but
should eliminate cases of non-leaf functions that don't have any
stack checking, which is a more serious problem.


git-svn-id: svn://svn.riscos.info/gccsdk/trunk@4656 5d7976f4-0b0e-0410-a586-d2bac8e3be52