- Fix for bug 214
* gcc/ira.c(ira_setup_eliminable_regset): When determining if
the frame pointer is needed, don't take alloca into account.
The use of alloca causes GCC to use the frame pointer (v6) to
access the stack, however, as we don't use the stack for alloca
v6 has not been initialised to point to sp. This patch ensures
that the frame pointer is eliminated into sp.
Note that in the GCCSDK 4.1 branch, this patch was part of
init_elim_table() in reload1.c. I believe that when migrating
to 4.6, the code had changed so much that it became unclear
where or if this patch should be applied.
* recipe/patches/gcc/unused/gcc.reload1.c.p: Delete old patch.
Code moved to gcc/ira.c.


git-svn-id: svn://svn.riscos.info/gccsdk/branches/developers@5049 5d7976f4-0b0e-0410-a586-d2bac8e3be52