Backport r5051 to trunk:
--8<--
* binutils/bfd/elf32-arm.c(riscos_module_reloccode): Add support for
undefined weak symbols in modules by changing the relocation code so
that symbols with a value of zero are not relocated.
Note that, as far as relocations are concerned, the start of the
module is defined as &8000 and Image$$RO$$Base as approximately &80A8
(and not 0 as might be assumed). Therefore any relocations that may
relate to these locations (however unlikely) will not be misinterpreted.
--8<--

gcc4/riscos/testsuite/module/004-weak: Add test.



git-svn-id: svn://svn.riscos.info/gccsdk/trunk/gcc4@5059 5d7976f4-0b0e-0410-a586-d2bac8e3be52