Merge r5143 from branches/developers/joty/gcc47 trunk to :
------------------------------------------------------------------------
r5143 | lee | 2011-05-02 21:18:37 +0200 (Mon, 02 May 2011) | 15 lines

* gcc/config/arm/arm.c(arm_emit_call_insn): Make sure the PIC
register is loaded for a call through the PLT.
(arm_expand_prologue): Make sure the PIC register is loaded
before the stack extension routines are called without using
require_pic_register(). This function causes an assertion
failure, probably because it is called too late in the process.
However, as our requirements are quite strict, ie, the PIC
register is known to be r7 and the code to load it is fixed,
we can use bits of require_pic_register() to achieve what we
want.
(arm_function_ok_for_sibcall): Use code similar to
TARGET_VXWORKS_RTP to prevent tail call optimisation via PLT
(PIC register is live on entry to PLT so the call must be made
before restoring the PIC register).



git-svn-id: svn://svn.riscos.info/gccsdk/branches/developers@5242 5d7976f4-0b0e-0410-a586-d2bac8e3be52