Sun Nov 23 14:53:00 GMT  2014  Lee Noar  <leenoar@sky.com>
	* pthread/cancel.c(__pthread_pull_gthr_weak): Remove. We
	now disable weak pthread symbols when building libstdc++
	as static.


Also:
* libstdc++-v3/config/os/riscos/ctype_base.h: New file copied from
os/generic.
* libstdc++-v3/config/os/riscos/ctype_configure_char.cc: Likewise.
* libstdc++-v3/config/os/riscos/ctype_inline.h: Likewise.
* libstdc++-v3/config/os/riscos/error_constants.h: New file copied
from os/generic with r4710 applied.
* libstdc++-v3/config/os/riscos/os_defines.h: Disable weak references
for pthread functions when building statically.
* libstdc++-v3/configure.host: Make sure os/riscos directory is found.


This issue was previously fixed in r3516, but the selection of pthread
functions pulled in by __pthread_pull_gthr_weak is no longer enough.
As we know that the pthread functions are always available, a better fix
is to disable the weak references. The build system provides a mechanism
for this (and some of the other targets use it as well).


Fix for bug#254.


git-svn-id: svn://svn.riscos.info/gccsdk/trunk@6748 5d7976f4-0b0e-0410-a586-d2bac8e3be52