Add SDL2 and SDL2-ttf.
These are targeted specifically at the RaspberryPi and make use of
hardware accelerated OpenGL via Khronos.
There is currently no sound support and it is single tasking only,
ie no RISC OS windows, but it does show what the OpenGL graphics
are capable of in terms of full screen games.


git-svn-id: svn://svn.riscos.info/gccsdk/trunk@7096 5d7976f4-0b0e-0410-a586-d2bac8e3be52